initial release
This is a web application written in Meteor to control a house
competition for the game ‘Family Feud’. In this variation, there are
four houses competing for points.

The default is the admin view. To turn this off for projecting on a
screen, enter Session.set(‘isAdmin’,false) in the console.